Add a rudimentary test framework; update for objasm.
The tool is intended here to be used for processing objasm files.
This means that I want to handle some more of the RISC OS specific
ideosyncracies. Ideally, I'd like to handle anything that ObjAsm
has, but I doubt that's possible.

Right now, the changes in the code have been related to the AREA
directive and the `=` literal string/byte sequence specifier. There's
also some small fixes for things like the hex digits and typos.