Ensure that expectations are kept; clean up when an error happens.
We now keep the expectations of the tests. This means that we'll be
able to see when regressions occur, which will make it easier to ensure
that we don't unintentionally break things.

When an error does occur in the building of files, we now delete the
partial file that was written. If we didn't, we'd end up with the file
being left behind and you couldn't rely on the tool in a makefile.