Apply some small speed improvements.
The simple hello_world_pp was taking 12 seconds to process; it's now down
to 9 seconds, but that's still way too long for processing the files.

The changes here have been to anchor some of the regexes and add some
better checks for the content forms so that we don't perform long ops
when we don't need to.

The inner processing loop probably needs to be reworked so that we don't
have to repeatedly match regexes - this was a lot easier in the Hdr2H
code as it didn't need to cope with some corner cases, but I think we
can repeat the same sort of processing here and get improved performance
in the future.