Initial support for conditionals.
Conditionals are going to have to be processed internally if they're
non-simple. The most basic of conditionals which just check for
defined values might be passed through, but in general this isn't
going to work for the other cases - we can't do string manipulation
in the GNU as assembler, and anything with booleans or relations
is not supported either. Other than the {TRUE} and {FALSE} cases,
we've got nothing useful we can really do here, so this is likely
to fail without internal processing.

The implementation here passes the ifdef tests but not the simple
{TRUE} case. We'll need more thought to support them.