Add support for some built in variables.
Built in variables must be able to be evaluated, especially within
conditionals, as they will commonly be used to check what the behaviour
should be for a specific architecture.

We now expand some of the built in variables into values that we can
evaluate sensibly. Not all of these will work in the ObjAsm sense, yet.