Improved the expression evaluation to include more operations.
The expression evaluation now includes some more operations to make
it better able to hanlde some expressions and conditionals. We're
still a long way off being able to handle objasm syntax directly,
but the expressions are better than they were. We do, however, still
rely on the regex parsing to hit expressions properly - it is quite
possible we'll replace something in the middle of a string by not
being careful. This could change in the future, and because we have
tests to check the behaviour we can see that things aren't broken.