Improved the expression parser to take over from the old evaluate.
The old evaluation function would work its way through the line text
performing substitutions and then try to evaluate (in perl) the result.
This was both unreliable in the fast of strings and did not handle many
of the cases. We now perform a much more stringent set of expression
operations to convert the line. This may be overkill but it is more
likely to interpret the files correctly.

It's still not used everywhere, so there will still be some
inconsistencies. It may go wrong in some cases too. But it largely
passes the tests - and fixes some that had got the wrong values
in them and hadn't been noticed.