Add support for local variables.
Local variables had been recognised, but were treated as just being
another class of global, because the stacking had not been performed.
Now we stack the variables on macro entry and unstack on exit so that
the locals come out ok. Tested with a simple macro expansion that
depends on the interpretation of the LCL.