Improve current PC by attempting partial expression evaluation.
When using {PC} or . we now try to evaluate partial expressions in
order to pass something to the assembler that it can understand.
This is still not ideal but it gives us a little bit better an
understanding of the expressions without breaking the code too
much when things go wrong.